protocol/bc: move SigHasher into its own file
